+++ changelog.md
@@ -2,6 +2,17 @@
 
 > All Time Stamps: GMT+8
 
+## 0.2.53.0 _2020-03-12 10:00_
+
+* 增加方块和物品 added blocks and items
+  * 东陵石矿 aventurine ore block
+  * 东陵石 aventurine item
+  * 东陵石块 aventurine block
+* 增加材料 added materials
+  * 东陵石 aventurine
+* 增加属性 added traits
+  * 精力汇聚 stamina focusing
+
 ## 0.2.52.0 _2020-03-11 15:00_
 
 * 增加状态效果 added potion effects